Refacor: 피드백 기반 수정
-불필요한 안전검사 및 로그 출력 제거
-ObjectLinker 함수 ID 부분 uint8로 수정

+++ Source/Endless_Hotel/GameSystem/Anomaly/Anomaly_Generator.cpp
@@ -49,7 +49,7 @@ void AAnomaly_Generator::AnomalyObjectLinker()
 	{
 		auto* Sub = GetGameInstance()->GetSubsystem<UAnomalyProgressSubSystem>();
 		auto* AnomalyObject = Cast<AAnomaly_Object_Base>(FoundActor);
-		int32 LinkObjectID = Sub->GetAnomalyDataByID(CurrentAnomaly->AnomalyID);
+		uint8 LinkObjectID = Sub->GetAnomalyDataByID(CurrentAnomaly->AnomalyID);
 		if(LinkObjectID == AnomalyObject->ObjectID)
 			CurrentAnomaly->LinkedObjects.Add(AnomalyObject);
 	}
